TODO: kludge to work around my OpenSSH workaround in .vimrc -- I set
'shellescape' there, which seems to make OpenSSH work better on Windows, #
Please enter the commit message for your changes. Lines starting but it also
causes shellescape to put *single* quotes around pattern     # with '#' will
be ignored, and an empty message aborts the commit.  here. I should fix the
problem with OpenSSH functionality instead of the # On branch nonminimal hack
in .vimrc, since it's apt to affect plugins besides this one.