WIP: move old non-packaged scripts (plus a resurrected part of a package I used to use) into "ATTIC"
This includes some old Rails- and Cucumber-related things that I may
need to revisit some time in the future; vimwiki stuff which I also
might want to revisit; and a.vim-related files that had been deleted
when I ostensibly switched to using vim-projectionist for their
functionality, which I might also want to sort through and see if they
might be useful in the future. I haven't actually explored
vim-projectionist yet, so I don't know if I'm missing anything that was
in a.vim.